feat: optimized select field
